Fddb: add RejectReadOnly to MysqlOptions
Over the weekend of 11th, 12th July an incident where a failover of database by
AWS resulted in errors on archimedes consumer due to a switch of a writer to a
reader. Archimedes consumer and other applications could not automatically
recover as expected during the incident.

This is because the mysql driver held the connection open with the read-only
replica. To ensure that this doesn't occur again, we want to enable the
https://github.com/go-sql-driver/mysql#rejectreadonly option. Add this option to
the MysqlOptions so that users of the library can add it if necessary.

https://jira.deliveryhero.com/browse/CART-3403

Signed-off-by: Karthik Nayak <karthik.nayak@deliveryhero.com>